cleaner solution than delays
