+++ app/src/androidTest/java/com/github/se/gatherspot/ui/EventsViewCompleteTest.kt
@@ -110,6 +110,7 @@ class EventsViewCompleteTest {
       eventRow.performClick()
     }
     ComposeScreen.onComposeScreen<EventUIScreen>(composeTestRule) {
+      composeTestRule.waitUntilAtLeastOneExists(hasTestTag("profileIndicator"), 10000)
       // Test the back button
       backButton {
         assertIsDisplayed()
@@ -121,6 +122,7 @@ class EventsViewCompleteTest {
       eventRow.performClick()
     }
     ComposeScreen.onComposeScreen<EventUIScreen>(composeTestRule) {
+      composeTestRule.waitUntilAtLeastOneExists(hasTestTag("profileIndicator"), 10000)
       // Test the profile indicator
       profileIndicator {
         assertIsDisplayed()
