+++ app/src/main/java/com/github/se/gatherspot/firebase/EventFirebaseConnection.kt
@@ -462,44 +462,7 @@ class EventFirebaseConnection : FirebaseConnectionInterface<Event> {
   }
 
   /*
-  fun cleanCollection() {
-    Firebase.firestore
-        .collection(EVENTS)
-        .whereNotEqualTo("organizerID", "")
-        .get()
-        .addOnSuccessListener { querySnapshot ->
-          Log.d(TAG, "Found ${querySnapshot.documents.size} documents with non empty organizerID")
-          querySnapshot.documents.forEach { document ->
-            Firebase.firestore
-                .collection("clean_events")
-                .document(document.id)
-                .set(document.data!!)
-                .addOnSuccessListener {
-                  Log.d(TAG, "DocumentSnapshot successfully moved to clean_events : ${document.id}")
-                }
-                .addOnFailureListener { e -> Log.w(TAG, "Error moving document", e) }
-          }
-        }
-        .addOnFailureListener { exception -> Log.d(TAG, exception.toString()) }
-  }
 
-
-
-  fun retrieveEvents() {
-    Firebase.firestore
-        .collection("clean_events")
-        .get()
-        .addOnSuccessListener { querySnapshot ->
-          Log.d(TAG, "Found ${querySnapshot.documents.size} documents in clean_events")
-          querySnapshot.documents.forEach { document ->
-            val event = getFromDocument(document)
-            if (event != null) {
-              add(event)
-            }
-          }
-        }
-        .addOnFailureListener { exception -> Log.d(TAG, exception.toString()) }
-  }
 
   fun retrieveMissing() {
     Firebase.firestore
